fix/QB-1896 RPC put use correct filename
